backend/scripts/requirements_demo.txt
psycopg2-binary
faker
bcrypt
reportlab
python-dotenv
